trusted-types/WorkerGlobalScope-importScripts.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
  //helper function for the tests
  function NewWorker(url, t) {
    var worker = new Worker(url);
    worker.onmessage = t.step_func_done(e => {
      assert_equals(e.data, "Pass");
    });
    worker.postMessage("ping");
  }

  async_test(t => {
    var url = "support/importScripts-from-trusted-script-URL.js";
    NewWorker(url, t);
  }, "importScripts from TrustedScriptURL unsafelyCreate.");
</script>
